+++ src/test/groovy/ru/vyarus/gradle/plugin/mkdocs/task/VersionsFileTest.groovy
@@ -364,6 +364,86 @@ class VersionsFileTest extends AbstractKitTest {
         result.output.contains('existing alias \'1.x\' override')
     }
 
+
+    def "Check version hiding"() {
+
+        setup:
+        initVersions('1.1', '1.2', '1.11', '1.21', '1.2-rc.1', '1.2-rc.2', '1.1-SNAPSHOT', '1.21.1')
+        build """
+            plugins {
+                id 'ru.vyarus.mkdocs'                                
+            }
+            
+            version = '1.22'
+            
+            mkdocs.publish.hideVersions '1.2-rc.1', '1.2-rc.2'
+            
+            python.scope = USER           
+        """
+
+        when: "run init"
+        BuildResult result = run('mkdocsInit')
+
+        then: "docs created"
+        result.task(':mkdocsInit').outcome == TaskOutcome.SUCCESS
+        file('src/doc/mkdocs.yml').exists()
+
+        when: "build versions"
+        result = run('mkdocsVersionsFile')
+        println 'project repo copy: ' + file('/.gradle/gh-pages/').list()
+
+        then: "published"
+        result.task(':mkdocsVersionsFile').outcome == TaskOutcome.SUCCESS
+        result.output.contains('hidden: 1.2-rc.2, 1.2-rc.1')
+
+        then: "versions file correct"
+        file('/build/mkdocs/versions.json').exists()
+        with(new JsonSlurper().parse(file('/build/mkdocs/versions.json')) as List) {
+            // yes, snapshots and rc go before released version - it's OK (because version format is unknown)
+            it.collect {it['version']} == ['1.22', '1.21.1', '1.21', '1.11', '1.2', '1.1-SNAPSHOT', '1.1']
+        }
+    }
+
+
+    def "Check auto bugfix versions hiding"() {
+
+        setup:
+        initVersions('1.1', '1.2', '1.11', '1.21', '1.2-rc.1', '1.2-rc.2', '1.1-SNAPSHOT', '1.21.1', '1.21.2', '1.22.1.1', '1.22.1.2')
+        build """
+            plugins {
+                id 'ru.vyarus.mkdocs'                                
+            }
+            
+            version = '1.23'
+            
+            mkdocs.publish.hideOldBugfixVersions = true
+            
+            python.scope = USER           
+        """
+
+        when: "run init"
+        BuildResult result = run('mkdocsInit')
+
+        then: "docs created"
+        result.task(':mkdocsInit').outcome == TaskOutcome.SUCCESS
+        file('src/doc/mkdocs.yml').exists()
+
+        when: "build versions"
+        result = run('mkdocsVersionsFile')
+        println 'project repo copy: ' + file('/.gradle/gh-pages/').list()
+
+        then: "published"
+        result.task(':mkdocsVersionsFile').outcome == TaskOutcome.SUCCESS
+        result.output.contains('hidden: 1.22.1.1, 1.21.1, 1.2-rc.1')
+
+        then: "versions file correct"
+        file('/build/mkdocs/versions.json').exists()
+        with(new JsonSlurper().parse(file('/build/mkdocs/versions.json')) as List) {
+            // yes, snapshots and rc go before released version - it's OK (because version format is unknown)
+            it.collect {it['version']} == ['1.23', '1.22.1.2', '1.21.2', '1.21', '1.11', '1.2-rc.2', '1.2', '1.1-SNAPSHOT', '1.1']
+        }
+    }
+
     private void initVersions(String... paths) {
         Grgit prjRepo = Grgit.init(dir: initDir)
         prjRepo.remote.add(name: 'origin', url: repoDir.canonicalPath, pushUrl: repoDir.canonicalPath)